requirements.txt
black==21.11b1
Django==3.2.9
djangorestframework==3.12.4
djlint==0.6.8
ipython==7.30.1
pyjwt==2.3.0
requests==2.26.0
waitress==2.0.0
